requirements.txt
groq
pandas
python-dotenv
pydantic
pytz
flask
requests
